pyproject.toml

[project]
name = "awscdk-python-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Erik Lundevall Zara", email = "45368867+eriklz@users.noreply.github.com"},
]
dependencies = [
    "aws-cdk-lib>=2.86.0",
    "constructs>=10.2.67",
    "tomli>=2.0.1",
    "structlog>=23.1.0",
    "typing-extensions>=4.7.1",
    "icecream>=2.1.3",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.3.0",
    "ruff>=0.0.275",
    "pre-commit>=3.3.3",
]
